feat(oprish): isolate model logic to its own module (#62)
* feat(todel): isolate module logic into logic module

* feat(todel): improve logic module position

* fix(effis): fix typo


* fix(todel): properly include files logic under http feature

---------
