+++ todel/src/models/logic/mod.rs
@@ -1,0 +1,5 @@
+#[cfg(feature = "http")]
+mod files;
+
+#[cfg(feature = "http")]
+pub use files::*;
